+++ source-code/cljs/plugins/item_lister/update/subs.cljs
@@ -29,7 +29,7 @@
   ; @param (map) server-response
   ;
   ; @example
-  ;  (r update.subs/server-response->deleted-item-ids :my-lister {my-handler/delete-items! ["my-item"]})
+  ;  (r update.subs/get-deleted-item-ids :my-lister {my-handler/delete-items! ["my-item"]})
   ;  =>
   ;  ["my-item"]
   ;
@@ -45,7 +45,7 @@
   ; @param (map) server-response
   ;
   ; @example
-  ;  (r update.subs/server-response->duplicated-item-ids :my-lister {my-handler/duplicate-items! [{:my-type/id "my-item"}]})
+  ;  (r update.subs/get-duplicated-item-ids :my-lister {my-handler/duplicate-items! [{:my-type/id "my-item"}]})
   ;  =>
   ;  ["my-item"]
   ;