+++ source-code/cljs/plugins/item_lister/update/validators.cljs
@@ -47,15 +47,3 @@
   (let [mutation-name    (r update.subs/get-mutation-name db lister-id :duplicate-items!)
         duplicated-items (get server-response (symbol mutation-name))]
        (vector/nonempty? duplicated-items)))
-
-(defn undo-duplicate-items-response-valid?
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  ;
-  ; @param (keyword) lister-id
-  ; @param (map) server-response
-  ;
-  ; @return (boolean)
-  [db [_ lister-id server-response]]
-  (let [mutation-name    (r update.subs/get-mutation-name db lister-id :undo-duplicate-items!)
-        deleted-item-ids (get server-response (symbol mutation-name))]
-       (vector/nonempty? deleted-item-ids)))